requirements.txt
asgiref==3.2.7
astroid==2.3.3
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.4.4
pytz==2019.3
sqlparse==0.3.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
